To dotty

+++ src/main/scala/org/bbstilson/raytracer/models/Translation.scala
@@ -9,7 +9,7 @@ case class Translation(m: Matrix) extends MatrixDouble(m) {
 
 object Translation {
   def apply(x: Int, y: Int, z: Int): Translation = {
-    val vals = Vector(x, y, z, 1)
+    val vals = Vector(x.toDouble, y.toDouble, z.toDouble, 1d)
     val f = (r: Int, c: Int) => if (r == c) 1d else if (c == 3) vals(r) else 0d
     val m = mkMatrix(4, 4, f)
     new Translation(m)
